backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
fastapi-jwt-auth==0.5.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pydantic==1.10.13
